% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeightedAverage.R
\name{calculateWeightedAverage}
\alias{calculateWeightedAverage}
\title{Calculate weighted average variant allele fraction (VAF)}
\usage{
calculateWeightedAverage(vaf)
}
\arguments{
\item{vaf}{GRanges or GRangesList with variants and VAFs}
}
\value{
weighted average VAF
}
\description{
Calculate weighted average variant allele fraction (VAF)
}
\examples{
calculateWeightedAverage(data_fusion$vaf)

}
