% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{countBreakpoints}
\alias{countBreakpoints}
\title{Count unique breakpoints in breakpoints GRangesList}
\usage{
countBreakpoints(breakpoints)
}
\arguments{
\item{breakpoints}{GRangesList}
}
\value{
GRangesList of breakpoints annotated with counts
}
\description{
Collapses breakpoints and adds metadata column with count for each
breakpoint
}
