% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_fusion}
\alias{data_fusion}
\title{QIAseq ES Example Data}
\format{
List of intermediate objects

\describe{
\item{\code{gr}}{GRanges from parseBAM}
\item{\code{breakpoints}}{Top breakpoint from detectBreakpoints}
\item{\code{genes}}{GRanges with candidate genes}
\item{\code{vaf}}{Breakpoint with VAF from calculateVAF}
\item{\code{supporting_reads}}{String list of supporting reads}
\item{\code{nonsupporting_reads}}{String list of nonsupporting reads}
}
}
\usage{
data_fusion
}
\description{
QIAseq ES Example Data
}
\examples{
data_fusion$gr
}
\keyword{datasets}
