% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectBreakpoints.R
\name{detectBreakpoints}
\alias{detectBreakpoints}
\title{Detect structural variants}
\usage{
detectBreakpoints(
  gr,
  genes,
  min_count = 1,
  top_n = Inf,
  detect_reciprocal = FALSE
)
}
\arguments{
\item{gr}{GRanges with reads}

\item{genes}{GRanges with one or two candidate genes}

\item{min_count}{Minimum number of supporting reads}

\item{top_n}{Top n variants to return}

\item{detect_reciprocal}{Detect reciprocal rearrangement for two candidate
genes (default = FALSE)}
}
\value{
GRangesList with detected variants, each defined by two breakends
}
\description{
Find location of breakpoints from split reads.
}
\details{
genes must be a GRanges with genomic coordinates of a single gene for
intragenic variants, or two genes for intergenic variants.

top_n specifies the number of variants to return

min_count filters variants by minimum number of supporting reads
}
\examples{
detectBreakpoints(data_fusion$gr, data_fusion$genes)
detectBreakpoints(data_fusion$gr, data_fusion$genes, 3, 1)

}
