% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBreakpoints.R
\name{extractBreakpoints}
\alias{extractBreakpoints}
\title{Extract breakpoints from aligned reads}
\usage{
extractBreakpoints(gr, genes)
}
\arguments{
\item{gr}{GRanges with reads}

\item{genes}{GRanges with one or two candidate genes}
}
\value{
GRangesList with breakpoints
}
\description{
Extract breakpoints from aligned reads
}
