% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extractNonSplitReads}
\alias{extractNonSplitReads}
\title{Extract non-split reads}
\usage{
extractNonSplitReads(gr)
}
\arguments{
\item{gr}{GRanges with reads}
}
\value{
List with R1, R2
}
\description{
Extract non-split reads
}
