% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extractSplitReads}
\alias{extractSplitReads}
\title{Extract split reads}
\usage{
extractSplitReads(gr)
}
\arguments{
\item{gr}{GRanges with reads}
}
\value{
List with R1, R2
}
\description{
Extract split reads
}
