% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNonSupportingReads.R
\name{getNonSupportingReads}
\alias{getNonSupportingReads}
\title{Get nonsupporting reads}
\usage{
getNonSupportingReads(
  gr,
  breakpoint,
  technology,
  min_overlap = 19,
  min_overlap_flanking = min_overlap,
  min_overlap_variant = min_overlap
)
}
\arguments{
\item{gr}{GRanges with reads}

\item{breakpoint}{breakpoint}

\item{technology}{Technology of reads ("qiaseq" | "other")}

\item{min_overlap}{minimum overlap for each side of breakend (default = 19)}

\item{min_overlap_flanking}{minimum overlap for flanking region
(set to min_overlap if not specified)}

\item{min_overlap_variant}{minimum overlap for variant region
(set to min_overlap if not specified)}
}
\value{
List of nonsupporting reads by breakpoint
}
\description{
This function is a wrapper for selectReads which sets a default value for
overlap sizes if not specified.
}
\examples{
getNonSupportingReads(data_fusion$gr, data_fusion$breakpoint,
"qiaseq")

}
