% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSupportingReads.R
\name{getSupportingReads}
\alias{getSupportingReads}
\title{Get supporting reads}
\usage{
getSupportingReads(
  gr,
  breakpoint,
  technology,
  min_overlap = 19,
  min_overlap_flanking = min_overlap,
  min_overlap_variant = min_overlap,
  split_only = FALSE
)
}
\arguments{
\item{gr}{GRanges with reads}

\item{breakpoint}{breakpoint}

\item{technology}{Technology of reads ("qiaseq" | "other")}

\item{min_overlap}{minimum overlap for each side of breakend (default = 19)}

\item{min_overlap_flanking}{minimum overlap for flanking region
(set to min_overlap if not specified)}

\item{min_overlap_variant}{minimum overlap for variant region
(set to min_overlap if not specified)}

\item{split_only}{get split supporting reads only}
}
\value{
List of supporting reads by breakpoint
}
\description{
This function is a wrapper for selectReads which sets a default value for
overlap sizes if not specified.
}
\examples{
getSupportingReads(data_fusion$gr, data_fusion$breakpoint, "qiaseq")

}
