% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getTopBreakpoints}
\alias{getTopBreakpoints}
\title{Get top breakpoints from countBreakpoints GRangesList}
\usage{
getTopBreakpoints(breakpoints, top_n, min_count)
}
\arguments{
\item{breakpoints}{GRangesList with breakpoints}

\item{top_n}{Top n breakpoints to return}

\item{min_count}{Minimum number of required reads}
}
\value{
GRangesList with top breakpoints
}
\description{
Get top breakpoints from countBreakpoints GRangesList
}
