% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBAM.R
\name{loadBAM}
\alias{loadBAM}
\title{Parse a BAM file}
\usage{
loadBAM(input_bam, bounds = GRanges())
}
\arguments{
\item{input_bam}{Input bam file}

\item{bounds}{GRanges with bounding ranges (optional)}
}
\value{
GRanges with reads
}
\description{
loadBAM takes an input BAM file and an optional bounding range to get
aligned read segments and metadata from the file.
}
\details{
The function removes reads with unmapped mates, reads marked as
duplicates, and alignments with mapping quality zero. In the output,
metadata columns R1 and R2 indicate whether alignments are for
sequence reads 1 or 2, respectively. The output only includes alignments
for seqlevels in GenomeInfoDb::standardChromosomes().
}
