% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapsRegions.R
\name{overlapsRegions}
\alias{overlapsRegions}
\title{Get reads that overlap regions of interest}
\usage{
overlapsRegions(
  R1,
  R2,
  region_1,
  region_2,
  min_overlap_1,
  min_overlap_2,
  supporting,
  max_overlap
)
}
\arguments{
\item{R1}{read 1}

\item{R2}{read 2}

\item{region_1}{region 1}

\item{region_2}{region 2}

\item{min_overlap_1}{overlap for region 1}

\item{min_overlap_2}{overlap for region 2}

\item{supporting}{TRUE for supporting reads, FALSE for nonsupporting reads}

\item{max_overlap}{maximum considered overlap}
}
\value{
list with overlapping reads
}
\description{
Get reads that overlap regions of interest
}
