% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectReads.R
\name{selectReads}
\alias{selectReads}
\title{Select supporting or nonsupporting reads}
\usage{
selectReads(
  gr,
  breakpoint,
  technology,
  supporting,
  min_overlap_flanking,
  min_overlap_variant,
  split_only = FALSE
)
}
\arguments{
\item{gr}{GRanges with reads}

\item{breakpoint}{breakpoint}

\item{technology}{Technology of reads ("qiaseq" | "other")}

\item{supporting}{TRUE for supporting reads, FALSE for nonsupporting reads}

\item{min_overlap_flanking}{minimum overlap for flanking region}

\item{min_overlap_variant}{minimum overlap for variant region}

\item{split_only}{select split supporting reads only (ignored for
supporting = FALSE)}
}
\value{
List of supporting or nonsupporting reads by breakpoint
}
\description{
Get the names of reads that are supporting or nonsupporting
for a query breakpoint.
}
\details{
The function returns either supporting or nonsupporting reads,
specified by the supporting boolean.

TODO: Refactor function to return both supporting and nonsupporting at once.

For technology = qiaseq, only reads in the R12 direction are returned.
R1 is adjacent to the gene specific primer, thus R1 and R2 map to the
flanking and variant region, respectively.

A value for the minimum overlap in flanking and variant regions is required.
}
\examples{
selectReads(data_fusion$gr, data_fusion$breakpoint, "qiaseq", TRUE,
19, 19)

}
