% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_breakpoint}
\alias{validate_breakpoint}
\title{Validate breakpoint GRanges}
\usage{
validate_breakpoint(breakpoint)
}
\arguments{
\item{breakpoint}{GRanges object}
}
\value{
void
}
\description{
Validate breakpoint GRanges
}
