% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_breakpoints}
\alias{validate_breakpoints}
\title{Validate breakpoints GRanges or GRangesList}
\usage{
validate_breakpoints(breakpoints)
}
\arguments{
\item{breakpoints}{GRanges or GRangesList object}
}
\value{
void
}
\description{
Validate breakpoints GRanges or GRangesList
}
