% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_genes}
\alias{validate_genes}
\title{Validate candidate genes}
\usage{
validate_genes(genes)
}
\arguments{
\item{genes}{GRanges with candidate genes}
}
\value{
void
}
\description{
Validate candidate genes
}
